+++ Sources/TDDKit/XCTAssertWillDeallocateInstance.swift
@@ -1,6 +1,6 @@
 import XCTest
 
-extension XCTestCase {
+public extension XCTestCase {
     /// Fail a test if the instance was not deallocted.
     ///
     /// Memory leaks occur when an object is not properly released from memory.
@@ -27,8 +27,8 @@ extension XCTestCase {
     ///     let spy = Spy(userLoginResult: userLoginResult)
     ///     let sut = LoginFlow(service: spy)
     ///
-    ///     expectWillDeallocate(for: sut, file: file, line: line)
-    ///     expectWillDeallocate(for: spy, file: file, line: line)
+    ///     XCTAssertWillDeallocate(for: sut, file: file, line: line)
+    ///     XCTAssertWillDeallocate(for: spy, file: file, line: line)
     ///
     ///     return (sut, spy)
     /// }
@@ -40,9 +40,23 @@ extension XCTestCase {
     ///   The default is the filename of the test case where you call this function.
     ///   - line: The line number where the failure occurs.
     ///   The default is the line number where you call this function.
-    public func expectWillDeallocate(instance: AnyObject, file: StaticString = #file, line: UInt = #line) {
+    func XCTAssertWillDeallocate(
+        instance: AnyObject,
+        _ message: @escaping @autoclosure () -> String = "",
+        file: StaticString = #file,
+        line: UInt = #line
+    ) {
         addTeardownBlock { [weak instance] in
-            XCTAssertNil(instance, "Should have been deallocated. Possible memory leak.", file: file, line: line)
+            if let instance {
+                let description = [
+                    "XCTAssertWillDeallocate failed: should have been deallocated \"\(String(describing: instance))\"",
+                    message()
+                ]
+                    .filter { !$0.isEmpty }
+                    .joined(separator: " - ")
+                let context: XCTSourceCodeContext = .init(location: .init(filePath: file, lineNumber: line))
+                self.record(.init(type: .assertionFailure, compactDescription: description, sourceCodeContext: context))
+            }
         }
     }
 }