+++ Sources/TDDKit/XCTCaptureError.swift
@@ -7,30 +7,32 @@ public extension XCTestCase {
     /// For example:
     /// ```
     /// func test_failingFetchX_fetchY_fails() async throws {
-    ///     let error = AnyError()
+    ///     let error = XCTError()
     ///     let (sut, _) = makeSUT(fetchXResult: .failure(error))
     ///
-    ///     let capturedError = await captureError(from: try await sut.fetchY())
+    ///     let capturedError = await XCTCaptureError(from: try await sut.fetchY())
     ///
-    ///     XCTAssertNotNil(error)
-    ///     XCTAssertEqual(capturedError as? AnyError, error)
+    ///     XCTCastAssertEqual(capturedError, error)
     /// }
     /// ```
     /// - Parameter block: The function call under test.
     /// - Returns: The error thrown from the block or nil if no error thrown.
-     func captureError<T>(
+     func XCTCaptureError<T>(
         from block: @autoclosure () async throws -> T,
+        _ message: @autoclosure () -> String = "",
         file: StaticString = #file,
         line: UInt = #line
-     ) async -> Error {
+     ) async -> Error? {
         do {
             _ = try await block()
-            XCTFail("Should have thrown an Error.", file: file, line: line)
-            return CaptureError.noErrorThrown
+            let description = ["XCTCaptureError failed: should have thrown an error", message()]
+                .filter { !$0.isEmpty }
+                .joined(separator: " - ")
+            let context: XCTSourceCodeContext = .init(location: .init(filePath: file, lineNumber: line))
+            record(.init(type: .assertionFailure, compactDescription: description, sourceCodeContext: context))
+            return nil
         } catch {
             return error
         }
     }
 }
-
-public enum CaptureError: Error { case noErrorThrown }